requirements.txt
# MedVision.ai - AI Visual Question Answering System
# Core ML/AI
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
open-clip-torch>=2.20.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
openai>=1.0.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
gunicorn>=21.2.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0
pydicom>=2.4.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
# Pin httpx to 0.24.1 for compatibility with Starlette TestClient
httpx==0.24.1
pytest-asyncio>=0.21.0
